+++ cuACS/Animal.h
@@ -8,7 +8,7 @@ class Animal: public Profile
     public:
         Animal(int id, string name, string animalType, string breed, int age, string gender, string color, string size,
                int aggression, int attachment, int obedience, string energy, bool isCrateTrained, bool isHypoallergenic,
-               bool isNeutered, int childrenComfort, int loudness, int cost, int costPerYear, int intelligence, int cleanliness);
+               bool isNeutered, int childrenComfort, int loudness, float cost, float costPerYear, int intelligence, int cleanliness);
 		Animal(vector<string> statendife);
 		int getId();
 		string getName();
@@ -28,8 +28,8 @@ class Animal: public Profile
         bool getIsNeutered();
         int getChildrenComfort();
         int getLoudness();
-        int getCost();
-        int getCostPerYear();
+        float getCost();
+        float getCostPerYear();
         int getIntelligence();
         int getCleanliness();
 
@@ -54,8 +54,8 @@ class Animal: public Profile
         bool isNeutered;
         int childrenComfort;
         int loudness;
-        int cost;
-        int costPerYear;
+        float cost;
+        float costPerYear;
         int intelligence;
         int cleanliness;
 };
